Split Booru.swift into separate files for each class.
